+++ resources/views/blog/partials/preview.blade.php
@@ -1,10 +1,10 @@
 <article class="blog-preview" itemprop="blogPost" itemscope itemtype="https://schema.org/BlogPosting">
     <header class="blog-preview__heading">
-        <h1 class="blog-preview__heading-title" itemprop="headline">
+        <h2 class="blog-preview__heading-title" itemprop="headline">
             <a href="{{$blog->url}}" class="blog-preview__heading-title-link">{{$blog->title}}</a>
-        </h1>
+        </h2>
         <p class="blog-preview__meta" title="{{$blog->publication_date_formatted}}">
-            <time itemprop="datePublished" content="{{$blog->publication_date_formatted_rfc3339}}">
+            <time itemprop="datePublished" datetime="{{$blog->publication_date_formatted_rfc3339}}">
                 {{$blog->publication_date_for_humans}}
             </time>
         </p>