+++ resources/views/pages/item.blade.php
@@ -15,7 +15,7 @@
             @if ($page->updated_at > $page->publication_date)
                 <footer class="well well-sm page__last-updated">
                     {{ trans('general.last-update') }}
-                    <time itemprop="lastReviewed" content="{{$page->updated_at_formatted_rfc3339}}">
+                    <time itemprop="lastReviewed" datetime="{{$page->updated_at_formatted_rfc3339}}">
                         {{$page->updated_at_formatted}}
                     </time>
                 </footer>